# add a benchmark for cosine distance and unroll ARM Neon simd