similar to #937 this one optimizes cosine distance.

before
```
[300.01 ms 302.41 ms 305.41 ms] Cosine(simd)
[302.51 ms 306.08 ms 311.26 ms] Cosine(simd) second rng seed
```

after
```
[86.878 ms 87.004 ms 87.263 ms] Cosine(simd)
[87.301 ms 102.65 ms 131.89 ms] Cosine(simd) second rng seed
```

Will add AVX unroll later today